Rename PermissionEvaluator and move it to security package
